+++ docs/CHANGELOG.rst
@@ -52,7 +52,7 @@ Added
   them to FASTQ
 * Add HEAT-Seq pipeline tools
 * Add HEAT-Seq workflow
-* Add ``create-geneset`` processor
+* Add ``create-geneset``, ``create-geneset-venn``  processors
 
 Fixed
 -----